Add equals and hashcode in its own test class
